dev/requirements.txt
Click==7.0
colorzero==1.1
Flask==1.1.1
gpiozero==1.5.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
paho-mqtt==1.4.0
pip==19.2.3
PyYAML==5.1.2
setuptools==41.2.0
Werkzeug==0.15.6
wheel==0.33.6
